Fix poo #28648
Add an exception for QAM CaaSP workflow in respect to autoyast.
This is required for the stability of our scenarios. The `generic`
autoyast test is using _too many_ needle tags, thus our systems
are prone to miss the `bios-boot` tag/needle when we deploy a
*10-node* CaaSP cluster.

The reason behind this is that we are using autoyast just to bring
the workers up to the defined state. Any error that happens in
between, is not really important to capture it _on time_. In any
case, if something breaks during the autoyast installation, we will
encounter it anyway in the end of the installation.

Simply put, risking the stability of our tests in order to save
a couple of minutes it's something we would like to avoid.

As a result, this PR removes the ability of consistantly scanning
for errors during the autoyast installation, but it keeps looking
for the only needle we care specifically.