Fix poo #28648
Removes the ability of consistantly scanning for errors during the
autoyast installation. It also removes the checking for the grub
splash screen. The test kicks-in only when the nodes have properly
booted into the console prompt.

Also it increases the timeout for matching the bootstrap needle
for 10 nodes `stack_update-velum-10-nodes-outdated` since it
some tests it takes more time.

Last but not least, we upload the `velum` logs, since they
are not properly collected by `supportconfig`.

In case of failure, the test timeout in worst case in 20 minutes.