Use registration for autoyast installations
For SLE15 default scenario is to register system and activate
repositories provided by SCC. Hence, we should use this as default for
autoyast installations as well. For obvious reasons we cannot put reg
code in the public github repo, so we have introduced mechanism for
variable expansion, which doesn't require support server.

Workers require upgrade in order to use this feature. To be on safe side
and not to break other autoyast installations, profile modification is
performed when AUTOYAST_PREPARE_PROFILE variable is set. Consequently,
test suites will have to have this variable, and ISO_1 can be removed
for most of them.

See [poo#28349](https://progress.opensuse.org/issues/28349).