Expand variables for aytests profiles in support server
We already have a mechanism when variables are replaced when file is
requested using cgi. However, it's tricky if we want to use job
variables as they are not available in cgi. As a first step we simply
replace all occurrences in the folder with profiles. As it's possible to
modify profiles on the worker directly we should use unified approach.

See [poo#28349](https://progress.opensuse.org/issues/28349).