LTP: Prevent LTP from breaking other tests
This isolates the LTP within main.pm by loading it inside an eval. The LTP
test runner often uses new os-autoinst features which are not present on older
workers. Because main.pm is loaded for all tests, this causes all tests to
break if there is something wrong with the load test code for LTP.