Upload ay profile when modified and if installation fails
At the moment it's very confusing which profile has been used for the
installation. What makes it worse is that we normally collect profile in
clone test module, which doesn't help when installation fails.
With this commit we save ay tests profile using hack and ulogs directory
and if installation fails using post_fail_hook. It covers most of the
scenarios. If installer will crash and end up in linux rc, we won't be
able to get profile from SUT, but may consider downloading it and then
uploading as a log file.

Done as a part of [poo#28349](https://progress.opensuse.org/issues/28349).