utils: handle_login: do not send spurious ret
Commit fa407950 reworked the code to make it more readable, but added
an additional else block sending return. RET needs only be sent on DMs
that have a user selection to be done, which is handled in the IF cases.

The additional RET causes issues in for example sddm, where RET triggers
a login without password, passing this on to the pam stack. During this
time, the password login field is insensitive to further input and
openQA does not wait for the field to turn active again before typing.