nw_ascs_install: Add support for multiple ARCHs
This adds support to the sles4sap/nw_ascs_install test so it
can handle NetWeaver's sources from multiple architectures to be
present on the shared volume passed via the NW variable. It will
attempt to cd to a get_var('ARCH') subdirectory for the correct
sources before starting the installation; this way with a properly
configured sources share, the same value for NW can be used on
tests over different architectures without needing to create
architecture-specific test suites.

Also sets NetWeaver profile and solution on the test system
before the installation and changes the value of SAPINST_SLP_MODE
to false during installation (required by unattended installation
of NW 7.42 and on)