Do not select desktop applications module where not required
For sles we should have only base system and system applications
enabled. If we DESKTOP set to gnome explicitly, then we enable it, or in
case of sled. After check of test suites we have, there is no a single
test suite which doesn't have DESKTOP variable explicitly defined and
doesn't have INSTALLONLY or EXIT_AFTER_START_INSTALL set to 1 in
functional job group.
So since now on we will have textmode installation by default for sles
and if this behavior doesn't fit, DESKTOP variable can be explicitly set
in the test suite.

See [poo#29589](https://progress.opensuse.org/issues/29589).