sles4sap: Add sap_server pattern SYSTEM_ROLE=textmode
This changes sles4sap/patterns test so it checks also for the
sap_server pattern. If the system was installed with system
role 'SLES for SAP Applications', then sap_server pattern should've
been installed and this will report a failure; however if the
system was installed with the 'textmode', will attempt to
install the pattern and inform that it was missing.

Fixes sles4sap/sapconf test when SYSTEM_ROLE=textmode, which
fails after changes introduced with:
https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/4249
and https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/4262