LTP: Disable reboot after install to avoid snapshot
boot_ltp is marked as a milestone so that we can revert to it during normal
testing. It is also ran after install_ltp as a sanity check to make sure the
system boots into a clean state after installing the LTP (because a lot of
tests depend on it, so we don't want every test to fail in boot_ltp). However
the qcow2 image published by install_ltp has become very large since boot_ltp
was marked as a milestone.

We don't want to remove boot_ltp's milstone, so instead we can remove boot_ltp
from the install_module. At least temporarily.

Also snapshots currently fail a lot, especially for VMs with larger amounts of
RAM in use.