Apply workaround for cnf with package in different module
After changes to default desktop, we now don't have wireshark available,
whereas it helps us to test cnf feature for SLE 15 that cnf is capable
to find packages even they are in not activated modules.
As a workaround we temporarily activate desktop applications module and
deactivate after that, including soft failure.

See fate#323424