LTP/install: Improve adding Workstation Extension addon/module
scc_we_enabled() was added in commit
3bf6935c ("Install ntfsprogs if is WE available as registered product.")
Back then it was called *after* trying to add modules via variables.

Then it was called *before* in commit
03b6c228 ("LTP/install: Rewrite dependencies")
which cause reporting an error (red box) on SLE 15, which is false positive.

The fix would be to put it back to the end, but as there isn't currently
WE available as registered product, remove scc_we_enabled().

+ replace 'and' boolean operators to '&&' (and is disaloved to use).