yast2_dns_server: fixup usage of systemctl
The systsmctl helper function is not applicable here, as we make use of
a bash construct piped to grep. Without the proper placements of brackets,
the return value is 0, even though grep did find what we expected, which in
turn means we should see retval 0.

In this occurence, skip using the systemctl helper function and revert back
to using assert_script_run.