Fix workaround of live-CD reboot failure handling
Fixes a regression introduced by 18a65b9b in combination with a649d1e6.
Whenever we reset the consoles we have to ensure that the system was really
rebooted so that the next `select_console` call finds a not logged in session
as is expected for non-activated consoles. In the case of the KDE
Live-Installer however the reboot requested by the installer does not work
(boo#993885) but we already reset the consoles causing the test to look for
a 'text-login' needle when we would just encounter a still active
'install-shell'. This commit fixes this by moving the
`handle_livecd_reboot_failure` call within the `power_action` command where
the actual reboot is requested but failing to reboot the system so that after
exiting the `power_action` call we ensured that the system really rebooted as
well as that the consoles are reset.

Verification run: http://lord.arch/tests/505

Related progress issue: https://progress.opensuse.org/issues/31480