Remove module deactivation from post-run hook
As a workaround for cnf we activate desktop-applications module, and
tried to deactivate it afterwards to bring the SUT to original state.
However, we have bigger problem in our flow. We use same VM image in
many tests, but that image shares same SCC credentials, so
activation/deactivation in one scenario will affect other scenarios.

For example, on aarch64 some test suite activated dev tools module, so
when we tried to deactivate desktop-applications, SCC correctly
responded that we should deactivate dev-tools module first, which is not
activated in this particular SUT. So, as activation still works fine, we
leave this soft-failure, but remove cleanup with module deactivation.

See [bsc#1079848](https://bugzilla.suse.com/show_bug.cgi?id=1079848).