Delete obsolete test module no_perl_bootloader
This test module was only used to verify that no perl-Bootloader component is
installed which is not the case anymore. Also I checked for SLE15 that the
package is not part of any module so also not installed. The probability of it
returning is unlikely. Let us scrap this test.