netwaver_ascs: Replace call to wait_still_screen
Replace call to wait_still_screen in netweaver_ascs_install with
a call to check_screen 'root-console'. After further tests, it seems
it's not as much as the content of the screen is changing after a call
to 'systemctl restart systemd-logind', but just a reset of the screen.

A call to check_screen 'root-console' to verify the screen coupled
with a call to select_console works to ensure the test it's still
in the apropiate console.