poo#30616: Log xfreerdp's debug output
This adds debug logging to `xfreerdp`. We need it to debug cases where
`xfreerdp` from no aparent reason disconnects (or is disconnected) from
Hyper-V host (see poo#30616). Creates one log per `xfreerdp` connection.

The resulting log is about 25 MB per 80 minute job. Cleaning cron job is
set on the storage.

If the logging is not useful to gather more information for poo#30616,
or when we have the information, we might remove this.

Validation run:
* http://assam.suse.cz/tests/600#step/bootloader_hyperv/37