Use "boot_hdd_image" in all places where applicable to cover arch specific quirks
Previously in many cases "boot/boot_to_desktop" was used to boot into an
existing installation whereas in many cases this would have never worked on
some architectures. To correct this and ensure we are stating the right
expectations in all occurences this commit tries to use "boot_hdd_image" where
applicable. In most cases it will actually just schedule
"boot/boot_to_desktop" but will handle the special cases as well.

The order of "boot_to_desktop" and "support_server/wait_support_server" has
been adapted to handle the aarch64 specific way as well as the
"HA_CLUSTER"-specific way as previously used in "load_cluster_boot" used by
HPC tests as well as "load_ha_cluster_tests" which is actually using the
inverse for no apparent reason.

The test module "uefi_bootmenu" might be scheduled additionally where
previously it was not, e.g. on aarch64. However it will not have an additional
effect as there is a early-return with no-op within "uefi_bootmenu" for that
case.