partitioning_raid: Adjust prep partitioning size to recent YaST requirements
Will change the prep partition size as shown in
https://openqa.suse.de/tests/1555276#step/partitioning_raid/11 , should
prevent the warning as shown in
https://openqa.suse.de/tests/1555276#step/partitioning_raid/225
apparently introduced in recent SLE15 builds.

I did not test this on ppc64le. The prep partition size was never adapted
since ppc64le tests were introduced with 7d412a36 in 2015 so I assume we
should also be fine for other products.