sapconf: Ignore timeout on developer's test run
This commit removes the soft fail if `script_run` timeouts while
running the developers' supplied test script, as its status and
output is checked in steps after that. No need to soft fail with
error code `undef` on timeouts when further information is gathered
and reported while checking the generated log.