Detect cpu stuck bsc#1052258 and collect logs
As result of investigation of
[poo#30085](https://progress.opensuse.org/issues/30085) we have
identified that problem is not directly with the test, but that's not
obvious from how it looks. That's the best thing I believe we can do to
point to the bug so it's visible. Seems that it's already fixed in newer
kernel versions, hence not much sense in investigation.