Delete redundant and wrong test flag comments
We should not try to duplicate os-autoinst documentation so this commit
deletes the template code comments about the meaning of the test flags in the
modules. In all cases the old 'important' flag which does not exist anymore
was wrongly stated and new flags have not been so not having the comment at
all should be better.