Fix snapper related tests by moving snapshot test data from /root to /etc
The products changed and /root is now a subvolume as well so a snapshot will
not include changes in there. Moving to /etc should prevent this problem as we
can expect / always be snapshotted. This would require needle updates
to all products where we run yast2_snapper as well as yast2_snapper_ncurses
so I opted for a backward compatible approach first and only doing the change
for the products which are affected by the change of the /root subvol.

Related needle PRs:
* SLE15: https://gitlab.suse.de/openqa/os-autoinst-needles-sles/merge_requests/789

Verification runs:
* SLE15: http://lord.arch/tests/665

Related progress issue: https://progress.opensuse.org/issues/34018