Adjust autoyast tests flow for s390x
We got ay upgrade tests running on zVM, but didn't work for zKVM yet.
We tried to connect in smart way to check errors during the second
stage, but even with sync using startshell boot option it didn't work
and we decided to keep it as it is for now.

We have checked that these changes do not break existing tests,
including zVM, and tried to use common code we use to handle reboot not
to fix same issues in multiple places.

See [poo#11922](https://progress.opensuse.org/issues/11922).