LTP/network: Remove $LTPROOT/testscripts/network.sh emulation
+ remove logs expecting emulation has been run.

This work was initially intended to add support for these network tests
which are being rewritten into new shell API (poo#34177).
But then I realized that it's actually remove network.sh emulation as
it's not needed even for running network tests outside openQA [1].

[1] https://lists.linux.it/pipermail/ltp/2018-April/007681.html

Emulation calls test_net.sh which sets network. But even without this
emulation network setup is done. The difference is that with emulation
(= calling test_net.sh before testing + set some variables) the setup is
being run again before each test, which is a bit slower, but it's
guaranteed that network setup is properly restored (in case of previous
test has bug and fail to restore it).