Fix docker test on SLE12 validation tests
docker is part of the containers-module. Our SLE12 test scenario
extra_tests_in_textmode runs based on an unregistered SLES installation so the
module is not immediately available. Adding the update repo which includes the
module solves the problem same as is already done in other places. As the
validation tests focus on the base system and not the modules I guess we can
go with this approach.

Alternatives would have been:
* O1: Just register the whole system and add the module for just one test
  module -> deregister again after the test module
* O2: Publish the hdd image from a test suite that registers the system and
  add a new test scenario like "extra_tests_registered" where we would
  schedule the docker module -> Con: Quite different test flow for SLE12 vs.
  SLE15 as well as openSUSE

Verification run: http://lord.arch/tests/755

Related progress issue: https://progress.opensuse.org/issues/33184