Improve handling bsc#1008493 in yast2_tftp module
Once in a while we may match log needle, if error is not shown
immediately. So, move soft-failure further, so we actually try to close
dialog, and if it didn't work, we check if error is shown.

Be aware that this issue doesn't appear often, and wait_still_screen
would also do the trick, but we don't to wait for no reason.

See [poo#34072](https://progress.opensuse.org/issues/34072).