Assert tty selected by default and softfail with text-login
On hyperv we see issued with our flow, as we may start entering login
name before we switched to the selected tty. Hence, default behavior,
should be to assert that we have switched to required tty first. To be
on safe side, we fallback to old behavior to identify cases where it
didn't work and analyze those.

See
[poo#32926](https://progress.opensuse.org/issues/32926).