Split out the setup of the serial device to reduce the need to schedule consoletest_setup
"consoletest_setup" is a bit of a mixed case: It conducts some steps necessary
for "console" tests but also some of the other test modules that just use the
console or even just an xterm window. This commit splits out the setup of the
serial device which is most likely the most important step so that this step
can be scheduled in more test suite flows where it is needed but
consoletest_setup would be too big.

See https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/4900
as a motivation to do this.