svirt: Implement BOOTFROM, reorganise boot devices
Previously on Xen we hard-set boot device to disk on poweroff. Now, we
set it in bootloader_svirt to expected default and we are expected to
change it wherever we need to during the test scenario (e.g. in
reboot_after_installation.pm after OS installation).

Also KVM via svirt is now supported. `BOOTFROM` variable is now
implemented on svirt.

Validation runs:
* jeos-main: http://nilgiri.suse.cz/tests/477
* xfs: http://nilgiri.suse.cz/tests/461
* ext4: http://nilgiri.suse.cz/tests/469
* jeos-fs_stress: http://nilgiri.suse.cz/tests/472
* lvm+RAID1: http://nilgiri.suse.cz/tests/473
* textmode: http://nilgiri.suse.cz/tests/474