Use grub to detect reboot
Normally in the runs we are able to match bios-boot needle, but in some
scenarios we are to slow and already get grub menu which appears
afterwards. In case of the ay test for error detection during second
stage, we fail to do so as still are in the loop of the first stage
which is wrong. So exit the loop if detected grub.

We still have to leave second grub menu detection because we may match
bios-boot and stuck in grub if it has no timeout. This is done only
because we may break a lot of tests if remove bios-boot needles.

See [poo#34654](https://progress.opensuse.org/issues/34654).