JeOS: Do not lose screen resolution on Hyper-V & Xen PV
On Hyper-V the default screen resolution is higher than the openQA
standard one, on Xen PV it's lower, so we have to make it permanent in
grub configuration. Formerly we were loosing the configuration because
we were rolling back to Initial snapshot which does not have it. Now we
rollback to a custom snapshot.

Fails here: https://openqa.suse.de/tests/1640722#step/snapper_jeos_cli/39
Validation runs:
* jeos-filesystem@svirt-hyperv: http://nilgiri.suse.cz/tests/499
* jeos-filesystem@svirt-xen-pv: http://nilgiri.suse.cz/tests/505