other_DE: Reset lightdm settings after installing a new DE
lightdm stores the last session used for login by name, not by refering
to 'default', so once a new DE is installed which changes the
xsession-default.desktop (via update-alternatives), this is not direcly
used by lightdm.

Reset the lightdm cached information to get the next login done using
the configured system default.

Additionally, clean up setting the default_wm in
/etc/sysconfig/windowmanager: this file has been deprecated after Leap
42.x.