Enable update repos on SLE 15 by default
For SLE 15 previously we had an issue that update repos were replaced by
proxy SCC to non-existing ones, as we don't sync them. Now all update
repos per module are mapped to the same mirror on OSD. So it's safe to
enable this behavior for SLE 15.

See [poo#34099](https://progress.opensuse.org/issues/34099).