Simplify tty login prompt handling with tty$nr-selected only
To ensure we are on the right console we look for tty$nr-selected and already
recorded a soft-fail trying to get rid of uses of "text-login" in this
context. As we checked that no soft-fail was reported anymore on text-login
matching after tty$nr-selected did *not* match we can remove the
extra-handling of text-login here with low risk.

Related progress issue: https://progress.opensuse.org/issues/32926