Unlock encrypted drives on zVM
As a part of [poo#35959](https://progress.opensuse.org/issues/35959) we
would like to introduce test suite where we activate existing partitions
and scenarios based on cryptlvm are perfect candidates for it.
This PR add possibility to boot with encrypted partitions on zVM.

After it's merged, scenario can be enabled on zVM, however we need to
introduce special WORKER_CLASS to chain job in the way that they are
executed on the same worker, and hence reuse same DASD devices.