Select last node in snaphots tree to verify difference screen
In yast2_snapper_ncurses test depending on the run we may have multiple
directories in the snapshot. In ncurses we use end key to navigate to
last file in the tree, but if there are multiple roots, we navigate to
the directory node where expected menu is not shown. When selecting last
directory and expand it, test will work in both setups.

See [poo#36069](https://progress.opensuse.org/issues/36069).