Create encrypted partition to verify disk activation on zVM
We have technical limitation with zVM, as those are bare metal machines.
To test encrypted partitions we need to run 2 jobs sequentially which
cannot be guaranteed if machine is used for other jobs.
To still cover this scenario on zVM, we create empty encrypted partition
which then we activate.

See [poo#35959](https://progress.opensuse.org/issues/35959).