Verify update-test* packages available for SLE12SP4
As planned, we extend test module to test packages availability and
location. Note, this is not applied for SLE15 as we don't have non BETA
repos available at the moment, so it's hard to check where to expect
them.
Also, not all of the listed packages are available on iso, so do not
test all of them.

See [poo#36451](https://progress.opensuse.org/issues/36451).