LTP: Introduce var to enable reboot after installation
It appears the LTP installation module on SLE12-SP4 x86_64 publishes a
corrupted disk image which can not be booted (or even mounted). This
reintroduces a reboot to the LTP installation, but only if LTP_INSTALL_REBOOT
is set to one. So we can see if it is really the LTP installation corrupting
the disk image or the publishing process.