Add tests for InfiniBand testing
Add a new load_baremetal_tests to products/sle/main.pm, which for now
only loads the InfiniBand testsuite (and sets up the barriers needed for
this).
Add tests/kernel/ib_tests.pm that allows to execute a testsuite for
InfiniBand testing. The testsuite is currently under development at SUSE.

Currently, the hardware required for these tests is not available for
openSUSE testing.

Signed-off-by: Michael Moese <mmoese@suse.de>