Add ipa test module for public cloud testing
This module run IPA tool against an ec2 instance.
Before the tests starts we currently check for the existence of the
given image and upload otherwise.
The upload part will be changed in future, when the mesh tool will
take this part.
Use awscli from pip, cause current package in tumbleweed is broken.
The result files get uploaded to openQA.