Optimize trup_install
So far, `trup_install` fails in case you are trying to install
a package which has been already available on the system.

As a result this PR makes sure that it will install only
the packages that are not present on the current status of the
system and it will also skip the others (pre-installed). In
case all the packages are pre-installed, no action will be
taken; function will happily return.

Furthermore, it will also print a friendly notification
about it.