Prevent PCM related package installations which are not supported
The patterns from PCM conflict with each other and do not really relate
to realistic customer scenarios and are not supported. See also
https://progress.opensuse.org/issues/37114 and the linked bugs for more
details and explanations.

This change effectively removes the PCM module and patterns from the
"allmodules" test scenarios for all versions of SLE as well as
architectures. With the aforementioned ticket it is planned to bring
back tests covering the PCM but explicitly so we should not be afraid
about a reduction in test coverage.

Related progress issue: https://progress.opensuse.org/issues/37531