memtest: End the test after mem check passed
Well, it turn's out that `Esc` restarts not shutdowns the system. And I
really don't see any reason to restart the system after the check
passed, so just ending the test, when we are done. (Btw, we never
restarted the system before, we just send `Esc` to restart it, but
turned the system off immediately.)

Also removed the Hyper-V specifix check as `TIMEOUT_SCALE` does that for
us.

Fails here:
* memtest@64bit: https://openqa.suse.de/tests/1784183
* memtest@svirt-hyperv: https://openqa.suse.de/tests/1784204

Validation:
* memtest@svirt-hyperv: http://nilgiri.suse.cz/tests/244