JeOS: Record machine-id
https://progress.opensuse.org/issues/37832

This just records `/etc/machine-id` in openQA job. I don't see a
reliable (and easy) way in openQA to make sure `machine-id` differs for
every installation/deployment. One has to manually check that the
`machine-id` differs among instances of the same image, and among build
of the same machine. And that it's set at all.

Validation runs:
* `jeos-main@64bit-virtio-vga`: http://nilgiri.suse.cz/tests/261#step/record_machine_id/4
* `jeos-main@svirt-xen-hvm`: http://nilgiri.suse.cz/tests/262#step/record_machine_id/4
* `jeos-main@svirt-xen-pv`: http://nilgiri.suse.cz/tests/263#step/record_machine_id/4