Fix console selection for IPMI
Investigation on failing calls to power_action showed that the console
selected in ib_tests.pm is wrom..
When using select_console 'root-ssh', the console is working fine - but
power_action cannot be executed successfully. This commit fixes the
behavior for shutdown in this test by using use_ssh_serial_console
instead, which disables the SOL-console first and configures the correct
serial device.

Signed-off-by: Michael Moese <mmoese@suse.de>