Keep systemd timer active on "system performance" scenarios
Currently we have one test variable 'SOFTFAIL_BSC1063638' which is
related to the "system_performance" openQA test scenario on o3 as well
as osd. On older product versions we would only trigger cron jobs to run
for example a "btrfs balance" process expecting that any subsequent run
will basically be a no-op and have no significant impact on system
performance. Also, if the load of the system would be high we would
regard this as soft-fail pointing to bsc#1063638 because the expectation
is that background maintenance tasks should not induce high load but be
handled on full idle-capacity. This commit changes the newly added for
triggering systemd timer but keep the timers enabled when we also want
to look for bsc#1063638.

Verification run: http://lord.arch/tests/1280#step/force_scheduled_tasks/8

Related progress issue: https://progress.opensuse.org/issues/38090