Ensure desktop runner is responsive after reboot to fix sporadic failures
After reboot and immediately after login the system might be under bigger load
so that the desktop runner is not as responsive yet. By using the same
approach as in the desktop_runner test module in the reboot module after login
we ensure that any next module can rely on the desktop runner to be
responsive.

Related progress issue: https://progress.opensuse.org/issues/30805