reboot_plasma5: Fix sporadic failures and cleanup
The duplicate screen check with less-than-default timeout in the logoutdialog
search is harmful as is causing sporadic failures and not providing benefit as
the next step is clicking on the same screen anyway.

As all currently supported products behave the same we do not need to keep any
old-version code around and simplify the test module a lot.

Verification run: http://lord.arch/tests/1605

Related progress issue: https://progress.opensuse.org/issues/37000