Don't check for missing needle while booting ARM
For some reason we assert that boot-firmware is seen, but there appears to be
no such needle. Also this code used to report a soft failure, however the
soft failure was removed, but not the assert screen.

In any case we should not assert anything here, we should just press
enter. Booting on ARM works the same as X86 since the QEMU rewrite and if
someone is not using the rewrite, then they will go down a different code path
anyway.