Disable serial-getty service to avoid pollution of serial console
During research of the issue we have identified that one of the sources
which pollutes serial output and interferes script_output calls are
login prompt messages generated by serial-getty service.
So, along with packagekit, we mask it for console tests and unmask
afterwards.

See [poo#30613](https://progress.opensuse.org/issues/30613).