Remove workaround for shutdown on KDE plasma5
The commit removes workaround for the sutdown flow that breaks the test
as it checks twice for the same needle sddm_shutdown_option_btn.

The problem was that after first match the test immediately checks for
the same needle and founds it as system a little bit slow, then it tries
to press the shutdown button again, but the appropriate popup has
already gone.

[poo#36040](https://progress.opensuse.org/issues/36040)