Allow to set shutdown timeout with DEBUG_SHUTDOWN
The commit adds ability to specify shutdown timeout by assigning the
value to DEBUG_SHUTDOWN parameter.

This allows to manipulate with the shutdown timeout without adding
changes to the code, e.g. to gather more logs or check if the issue with
shutdown related to the low timeout.

Related ticket: [poo#35215](https://progress.opensuse.org/issues/35215)