Fix reboot_gnome on s390
Previously test checked for 'shut off' in case of s390 reboot_gnome, but
in fact, the vm state was 'running', as vm parameters were redefined
during the initial reboot and reboot=>destroy was changed to
reboot=>undef.

The commit adds an additional verification if the reboot is performing
on s390 zKVM system and it is an initial one (first_reboot). Then only in
that case call assert_shutdown_and_restore_system.

Related ticket: [poo#36279](https://progress.opensuse.org/issues/36279)