Install new package for CaaSP if any
* 'qam_packages.txt' contains all the packages included in the
   maintenance incident. e.g. [open-iscsi, iscsiuio]

* 'not_installed.txt' contains all the packages coming from
   the 'qam_packages' which are not pre-installed in the SUT.
   e.g. [open-iscsi]

We disable the update repository and search for all the packages
included in the 'not_installed.txt'. If we find any, this means
that these packages were not previously released to the customers.

* 'new_package.txt' contains all the packages that were not
  previously released to the customers. e.g. iscsiuio

If such packages exists, then we enable the testing repository and
we installed them. Afterwards, we parse again and we create new
files. We check again and if there is no 'not_installed' file
this means that all packages are ready to be update. Otherwise,
we disable the testing repo and we install them first. This is
what 'prepare' step stands for.

Last step, we check if this update included only new packages
or not. If yes, these packages are already installed in previous
step, so we skip the update. Otherwise, we proceed with updating
the remaining packages from the testing repository.