zypper migration retry on HTTP error 502
Sometimes zypper migration fails on Hyper-V with a following error:
https://openqa.suse.de/tests/2001946#step/zypper_migration/4. I was
wondering if we could retry couple of times to see if we eventually go
thru.

Currently the HTTP 502 error is not happening, so I can't test if it's
effective, but on the other hand it should not break anything and we
could find out on OSD if it works.