Hyper-V support for X11-to-VT switch
Implements the X11-to-VT switch itself: On Hyper-V console switch from
one console to another is not possible with the general 'Ctrl-Alt-Fx'
binding as the combination is lost somewhere in VNC-to-RDP translation.
For VT-to-VT switch we use 'Alt-Fx' binding, however this is not enough
for X11-to-VT switch, which requires the missing 'Ctrl'. However, in X11
we can use `chvt` command to do the switch for us.

I don't see in openQA way to know, if perticular console was activated
nor what console are currently at, hence the `CONSOLE_JUST_ACTIVATED`
test variable and the `check_screen()`, respectively. Though, looking
fragile, so far it was reliable.

Extracts code which maps console name to VT number to `console_nr()`.