Add SLES4SAP migration tests based on SLES code
This commit includes changes to SLES test code to make it possible
to reuse while testing offline and scc based upgrades with SLES for
SAP Applications. Changes are basically:

- Skip SLES4SAP product installation test during install when testing
upgrades.
- Skip btrfs snapshots tests from SLES on SLES4SAP.
- Do not attempt to log in as an unprivileged user in Gnome, as
SLES4SAP installation only sets up the system administrator, and the
SAP software installers set up different unprivileged users as the
ones defined elsewhere in the codebase.
- Check for extra needles in installation/disable_grub_timeout, as
even though bootloader configuration should not be available during
upgrades, it is available when upgrading SLES4SAP.
- Change SLES4SAP specific tests to also work with upgraded systems;
in particular patterns test does not require to fail if the sap_server
pattern is not installed on an upgraded system and sapconf test needs
to verify if the active profile is saptune in order to skip some parts
of the test.