Detect bsc#1063638 on btrfs
It's a bit tricky as we run btrfs maintenance tasks in the tests, so not
to have false positives we can put it after btrfs balancing is done, and
for other scenarios boot_to_desktop is the most common place, as for x11
tests we won't run consoletest_setup, and consoletest_setup has
force_scheduled_tasks after it, leading to the false positives.
Therefore, I would keep it that conservative and address scenarios which
are affected individually (e.g. yast_gui tests).

See [poo#39059](https://progress.opensuse.org/issues/39059).