yast2_dns_server: Accept current behaviour of slow startup (boo#1093029)
See https://bugzilla.suse.com/show_bug.cgi?id=1093029 for details. The test
code is expecting the systemd service to be running as soon as we exit the yast
configuration part but according to the YaST development team they see it as
acceptable when the service takes some seconds to start.