Expand kernel error checking to baremetal tests
Previously, LTP and InfiniBand tests did different kind of checking for
kernel errors, which should be unified.
This commit adds the checking for kernel errors, which is already
available for LTP tests, to the baremetal tests as well and removes the
duplicated functionality from ib_tests.
In addition, some more kernel errors will be checked for.

Signed-off-by: Michael Moese <mmoese@suse.de>