[CLOUD] Run LTP via SSH in public cloud
Use the script run_ltp_ssh.pl to run LTP tests on public cloud
instance. This is a interim solution and will be replaced, when a
proper backend for publiccloud testing is in place.
The script is using perl-expect to simulate the SSH connection
and setup commands to install and run LTP tests.
The results can be stored in json format, which will be uploaded
using parse_extra_log(LTP => file).

Created a base test class for publiccloud. Also used in
upload_image.pm.